Consolidate field accessors: merge primitive types and reduce code complexity (#6460)
apache/incubator-kie-issues#2106

Co-author: Claude

+++ drools-traits/src/test/java/org/drools/traits/compiler/factmodel/traits/StudentProxyImpl3.java
@@ -116,7 +116,7 @@ public class StudentProxyImpl3 extends TraitProxyImpl implements IStudent {
     }
 
     public void setD(double d) {
-        bit_writer.setDoubleValue( object, d );
+        bit_writer.setDecimalValue( object, d );
     }
 
     // get/set impl depends on bitmask
